Avoid import cycle
