+++ packages/liveblocks-react/src/__tests__/selectThreads.test.ts
@@ -1,7 +1,6 @@
 import type { ThreadData } from "@liveblocks/core";
 
-import { selectThreads } from "../liveblocks";
-import { UmbrellaStore } from "../umbrella-store";
+import { selectThreads, UmbrellaStore } from "../umbrella-store";
 
 describe("selectThreads", () => {
   it("should only return resolved threads from a list of threads", () => {
