+++ packages/liveblocks-react/src/liveblocks.tsx
@@ -33,8 +33,6 @@ import { useSyncExternalStore } from "use-sync-external-store/shim/index.js";
 import { useSyncExternalStoreWithSelector } from "use-sync-external-store/shim/with-selector.js";
 
 import { useIsInsideRoom } from "./contexts";
-import { byFirstCreated, byMostRecentlyUpdated } from "./lib/compare";
-import { makeThreadsFilter } from "./lib/querying";
 import { shallow2 } from "./lib/shallow2";
 import { useInitial, useInitialUnlessFunction } from "./lib/use-initial";
 import { use } from "./lib/use-polyfill";
@@ -46,13 +44,11 @@ import type {
   SharedContextBundle,
   ThreadsAsyncResult,
   ThreadsAsyncSuccess,
-  ThreadsQuery,
   UnreadInboxNotificationsCountAsyncResult,
   UserAsyncResult,
   UserAsyncSuccess,
   UseUserThreadsOptions,
 } from "./types";
-import type { UmbrellaStoreState } from "./umbrella-store";
 import { UmbrellaStore } from "./umbrella-store";
 
 /**
@@ -178,39 +174,6 @@ function selectorFor_useRoomInfo(
     isLoading: false,
     info: state.data,
   };
-}
-
-/**
- * @private Do not rely on this internal API.
- */
-// TODO This helper should ideally not have to be exposed at the package level!
-// TODO It's currently used by react-lexical though.
-export function selectThreads<M extends BaseMetadata>(
-  state: UmbrellaStoreState<M>,
-  options: {
-    roomId: string | null;
-    query?: ThreadsQuery<M>;
-    orderBy:
-      | "age" // = default
-      | "last-update";
-  }
-): ThreadData<M>[] {
-  let threads = state.threads;
-
-  if (options.roomId !== null) {
-    threads = threads.filter((thread) => thread.roomId === options.roomId);
-  }
-
-  // Third filter pass: select only threads matching query filter
-  const query = options.query;
-  if (query) {
-    threads = threads.filter(makeThreadsFilter<M>(query));
-  }
-
-  // Sort threads by creation date (oldest first)
-  return threads.sort(
-    options.orderBy === "last-update" ? byMostRecentlyUpdated : byFirstCreated
-  );
 }
 
 function getOrCreateContextBundle<
